Move scoped_annotation to separate header
